bug(refactor): refactored all file names to standard [finishes #165407939]
