+++ API/app/middlewares/Validator.js
@@ -1,7 +1,7 @@
 import Joi from 'joi';
-import User from '../models/users_model';
-import Account from '../models/Account_model';
-import Transaction from '../models/transaction';
+import User from '../models/User';
+import Account from '../models/Account';
+import Transaction from '../models/Transaction';
 
 class Validator {
     static SignUpValidator(req, res, next) {
