+++ API/utils/GenAccountNumber.js
@@ -1,4 +1,4 @@
-const AccountNumber = () => {
+const GenAccountNumber = () => {
     return (Math.floor(1000000000 + Math.random() * 9000000000));
 }
-export default AccountNumber;
+export default GenAccountNumber;
